유도미사일

+++ Dungreed/Utils.cpp
@@ -42,9 +42,9 @@ namespace MY_UTIL
 		float x = endX - startX;
 		float y = endY - startY;
 		float d = sqrt(x * x + y * y);
-		float angle = acos(x / d);
+		float angle = acosf(x / d);
 
-		if (y > 0)angle = PI_2 - angle;
+		if (y > 0) angle = PI_2 - angle;
 
 		return angle;
 	}
